+++ scripts/build_native_backend.sh
@@ -20,10 +20,11 @@ cmake_minimum_required(VERSION 3.22.1)
 project(hoshitan_backend LANGUAGES C CXX)
 
 add_subdirectory("$SRC_DIR" hoshidicts-build)
+find_package(SQLite3 REQUIRED)
 add_executable(iina-hoshi-dicts "$ROOT/src/native/iina_hoshi.cpp")
 set_property(TARGET iina-hoshi-dicts PROPERTY CXX_STANDARD 23)
 set_property(TARGET iina-hoshi-dicts PROPERTY CXX_STANDARD_REQUIRED ON)
-target_link_libraries(iina-hoshi-dicts PRIVATE hoshidicts)
+target_link_libraries(iina-hoshi-dicts PRIVATE hoshidicts SQLite3::SQLite3)
 CMAKEEOF
 
 cmake -S "$WRAPPER_DIR" -B "$BUILD_DIR" -DCMAKE_BUILD_TYPE=Release